fix: fix document title signal handling
When clicking on a side nav link, an error related to the document title
signal occurs:
Warning: Cannot update a component ('value') while rendering a different
component ('Layout').
This change updates the document signal handling to prevent the error.
The code is aligned with sources generated by start.vaadin.com.